% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_analysis.R
\name{mult_analysis}
\alias{mult_analysis}
\title{Multivariate analysis}
\usage{
mult_analysis(data, p_threshold = 0.05)
}
\arguments{
\item{data}{A data frame.}

\item{p_threshold}{Numeric variable between 0 and 1.}
}
\value{
The summary for the final logistic model based on FRACTURE outcome versus the significant numerical variables.
}
\description{
Fit the multivariable model comprising all variables identified in
the univariate analysis (univ_analysis()).
Then, the variables that do not contribute to the model
based in a p.value threshold (e.g., p-value > 0.05)
will be excluded and a new smaller model will be fitted.
}
\examples{
univ_analysis()

}
